@@ -22,113 +22,59 @@ namespace Voxed.WebApp
 
         public void ConfigureServices(IServiceCollection services)
         {
-            // Cors Configuration must be before MVC / Razor Configuration
-            services.AddCors(options =>
-            {
-                options.AddPolicy(name: myAllowSpecificOrigins,
-                                  builder =>
-                                  {
-                                      //builder.WithOrigins("http://localhost",
-                                      //                    "http://www.contoso.com");
-
-                                      builder.SetIsOriginAllowed(origin => new Uri(origin).Host == "localhost")
-                                      .AllowAnyHeader()
-                                      .AllowAnyMethod();
-                                  });
-            });
-
-            services.AddControllersWithViews();
-            services.AddRazorPages();
-
-            services.AddDistributedMemoryCache();
-            services.AddSession();
-            services.AddScoped<TraceIPAttribute>();
-
-
             // New approach
+            services.RegisterWebServices(_configuration);
             services.RegisterInfrastructureServices(_configuration);
             services.RegisterRepositories();
             services.RegisterServices(_configuration);
             services.RegisterIdentity();
-
-            services.ConfigureApplicationCookie(options =>
-            {
-                // Cookie settings
-                //options.Cookie.HttpOnly = true;
-                options.ExpireTimeSpan = TimeSpan.FromDays(30);
-
-                options.LoginPath = "/Identity/Account/Login";
-                options.AccessDeniedPath = "/Identity/Account/AccessDenied";
-                options.SlidingExpiration = true;
-            });
-
-            //Por default necesita que estes autenticado para entrar a los controllers 
-            //services.AddAuthorization(options =>
-            //{
-            //    options.FallbackPolicy = new AuthorizationPolicyBuilder()
-            //        .RequireAuthenticatedUser()
-            //        .Build();
-            //});
-
-            services.AddSignalR();
         }
 
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            // Elastic Apm Configuration must be before all configuration
-            //app.UseAllElasticApm(Configuration);
+        //public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        //{
+        //    app.UseCors(myAllowSpecificOrigins);
+
+        //    if (env.IsDevelopment())
+        //    {
+        //        app.UseDeveloperExceptionPage();
+        //    }
+        //    else
+        //    {
+        //        app.UseExceptionHandler("/Home/Error");
+        //        // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+        //        app.UseHsts();
+        //    }
+        //    app.UseHttpsRedirection();
+        //    //app.UseStaticFiles();
+        //    var cacheMaxAgeOneWeek = (60 * 60 * 24 * 7).ToString();
+        //    app.UseStaticFiles(new StaticFileOptions
+        //    {
+        //        OnPrepareResponse = ctx =>
+        //        {
+        //            ctx.Context.Response.Headers.Append(
+        //                 "Cache-Control", $"public, max-age={cacheMaxAgeOneWeek}");
+        //        }
+        //    });
+
+        //    app.UseRouting();
+
+        //    app.UseAuthentication();
+        //    app.UseAuthorization();
+
+        //    app.UseSession();
+
+        //    app.UseEndpoints(endpoints =>
+        //    {
+        //        endpoints.MapControllerRoute(
+        //            name: "default",
+        //            pattern: "{controller=Home}/{action=Index}/{id?}");
+
+        //        endpoints.MapRazorPages();
+        //        //endpoints.MapBlazorHub();
 
 
-            //app.UseCors(builder => builder
-            //    .AllowAnyOrigin()
-            //    .AllowAnyMethod()
-            //    //.AllowCredentials()
-            //    .AllowAnyHeader()
-            //    .SetPreflightMaxAge(TimeSpan.FromDays(365)));
-
-            app.UseCors(myAllowSpecificOrigins);
-
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-            app.UseHttpsRedirection();
-            //app.UseStaticFiles();
-            var cacheMaxAgeOneWeek = (60 * 60 * 24 * 7).ToString();
-            app.UseStaticFiles(new StaticFileOptions
-            {
-                OnPrepareResponse = ctx =>
-                {
-                    ctx.Context.Response.Headers.Append(
-                         "Cache-Control", $"public, max-age={cacheMaxAgeOneWeek}");
-                }
-            });
-
-            app.UseRouting();
-
-            app.UseAuthentication();
-            app.UseAuthorization();
-
-            app.UseSession();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllerRoute(
-                    name: "default",
-                    pattern: "{controller=Home}/{action=Index}/{id?}");
-
-                endpoints.MapRazorPages();
-                //endpoints.MapBlazorHub();
-
-
-                endpoints.MapHub<VoxedHub>("/hubs/notifications");
-            });
-        }
+        //        endpoints.MapHub<VoxedHub>("/hubs/notifications");
+        //    });
+        //}
     }
 }
